pyproject.toml
[tool.poetry]
name = "sqr-anime-voiceover"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
pandas = "^2.2.2"
uvicorn = "^0.29.0"
pytest = "^8.1.1"
google-api-python-client = "^2.127.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
openpyxl = "^3.1.2"
streamlit = "^1.33.0"
selenium = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
